Add the code RussellB to Valerino's run_mess.sh
Explanation of RussellB : 
The first line of code there takes the third-to-last parameter sent to the run_mess.sh and assumes it's the name of the rom. 
Next line gets the root of the filename minus full path and extension. 
The last remaining line of code replaces tmpmess.cmd with the root name of the rom and .cmd. 
That way it's unique per rom. 
This allowed me to specify custom configs, including bezels and screen locations, etc. per rom.