add shader dependency, path and help
Normally the shaders from the master branch are not installed on a rpi.
Though the gb-pocket-shader will work on the rpi4. (others not tested)
To get it working these shaders are cloned into the shaders-extra folder.
Now the can be used without interfering with the standard shaders installed on a rpi.